Move function to humanize numbers to data_helpers
